Implement access checking for Field DataProducer
In order to protect fields like first name and last name we must
implement access checking. The best place to do this is on the fields
themselves.

The `PropertyPath` data producer included in the GraphQL module doesn't
perform access checking so this should not be used for sensitive fields.

To ensure access is handled the same whether the GraphQL API is used or
a field is shown directly through Drupal the GraphQL API shouldn't
handle access directly. That's why the Field DataProducer defers to
field access checking.